linux: Move `if_packet.h` API to `src/new`
Link: https://git.kernel.org/pub/scm/linux/kernel/git/torvalds/linux.git/tree/include/uapi/linux/if_packet.h?h=028ef9c96e96197026887c0f092424679298aae8
Link: https://github.com/kernkonzept/l4re-core/blob/6d4b236ae59f69008a91176063851a8429bff22b/libc/uclibc-ng/contrib/uclibc/include/netpacket/packet.h#L35
