+++ src/new/l4re/mod.rs
@@ -1,3 +1,5 @@
 //! L4re.
 //!
 //! * Headers: <https://github.com/kernkonzept/l4re-core/tree/master/libc/uclibc-ng>
+
+pub(crate) mod packet;
